+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
+- Replaced `insane` with `@vtex/insane`.
 - Removed `ramda` calls.
 
 ## [0.10.1] - 2020-07-09